backend/requirements.txt
# FastAPI and server
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-multipart==0.0.9

# Database
sqlalchemy==2.0.25
aiosqlite==0.19.0

# Validation
pydantic==2.6.1
pydantic-settings==2.1.0

# HTTP client for external APIs
httpx==0.26.0

# Google Gemini AI
google-generativeai==0.4.0

# CORS
python-jose[cryptography]==3.3.0

# Environment variables
python-dotenv==1.0.1

# Date handling
python-dateutil==2.8.2
